# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProofPilot decision intelligence API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: decisions
    description: Human-agent decision workspaces
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /decisions:
    get:
      operationId: listDecisions
      tags: [decisions]
      summary: List decision workspaces
      responses:
        "200":
          description: Decision summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DecisionSummary"
    post:
      operationId: createDecision
      tags: [decisions]
      summary: Create a decision workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionInput"
      responses:
        "201":
          description: Created decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionState"
  /decisions/{decisionId}:
    get:
      operationId: getDecision
      tags: [decisions]
      summary: Retrieve the complete active decision state
      parameters:
        - $ref: "#/components/parameters/DecisionId"
      responses:
        "200":
          description: Complete decision state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionState"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateDecision
      tags: [decisions]
      summary: Update human-owned decision settings
      parameters:
        - $ref: "#/components/parameters/DecisionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionUpdate"
      responses:
        "200":
          description: Updated decision state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionState"
  /decisions/{decisionId}/analyze:
    post:
      operationId: analyzeDecision
      tags: [decisions]
      summary: Recalculate scores and findings from current decision state
      parameters:
        - $ref: "#/components/parameters/DecisionId"
      responses:
        "200":
          description: Recalculated state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionState"
  /decisions/{decisionId}/evidence:
    post:
      operationId: addEvidence
      tags: [decisions]
      summary: Add a source-backed evidence item
      parameters:
        - $ref: "#/components/parameters/DecisionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceInput"
      responses:
        "201":
          description: Added evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
  /decisions/{decisionId}/activity:
    get:
      operationId: getDecisionActivity
      tags: [decisions]
      summary: Retrieve the human-agent audit timeline
      parameters:
        - $ref: "#/components/parameters/DecisionId"
      responses:
        "200":
          description: Timeline entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEntry"
  /decisions/{decisionId}/sensitivity:
    get:
      operationId: getSensitivityAnalysis
      tags: [decisions]
      summary: Retrieve recommendation stability across changing priorities
      parameters:
        - $ref: "#/components/parameters/DecisionId"
      responses:
        "200":
          description: Sensitivity analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensitivityAnalysis"
  /decisions/{decisionId}/actions/weight-proposal:
    post:
      operationId: proposeWeightChange
      tags: [decisions]
      summary: Create a pending agent proposal to change a criterion weight
      parameters:
        - $ref: "#/components/parameters/DecisionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightProposalInput"
      responses:
        "201":
          description: Pending action requiring human review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingAction"
  /decisions/{decisionId}/actions/{actionId}/resolve:
    post:
      operationId: resolvePendingAction
      tags: [decisions]
      summary: Approve or reject a consequential agent proposal
      parameters:
        - $ref: "#/components/parameters/DecisionId"
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionResolutionInput"
      responses:
        "200":
          description: Resolved action and resulting state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionState"
  /decisions/{decisionId}/brief:
    post:
      operationId: generateDecisionBrief
      tags: [decisions]
      summary: Generate an explainable decision brief from current state
      parameters:
        - $ref: "#/components/parameters/DecisionId"
      responses:
        "200":
          description: Decision brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionBrief"
components:
  parameters:
    DecisionId:
      name: decisionId
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Decision not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
    DecisionInput:
      type: object
      required: [title, description]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        owner: { type: string }
        deadline: { type: ["string", "null"] }
    DecisionUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        status: { type: string }
        criteriaWeights:
          type: object
          additionalProperties:
            type: number
    DecisionSummary:
      type: object
      required: [id, title, status, recommendation, confidence, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        status: { type: string }
        recommendation: { type: string }
        confidence: { type: number }
        updatedAt: { type: string, format: date-time }
    DecisionState:
      type: object
      required: [id, title, description, status, options, criteria, evidence, findings, recommendation, pendingActions, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        owner: { type: string }
        status: { type: string }
        updatedAt: { type: string, format: date-time }
        options:
          type: array
          items: { $ref: "#/components/schemas/DecisionOption" }
        criteria:
          type: array
          items: { $ref: "#/components/schemas/DecisionCriterion" }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/Evidence" }
        findings:
          type: array
          items: { $ref: "#/components/schemas/Finding" }
        recommendation:
          $ref: "#/components/schemas/Recommendation"
        pendingActions:
          type: array
          items: { $ref: "#/components/schemas/PendingAction" }
    DecisionOption:
      type: object
      required: [id, name, shortName, description, scores]
      properties:
        id: { type: string }
        name: { type: string }
        shortName: { type: string }
        description: { type: string }
        website: { type: ["string", "null"] }
        scores:
          type: object
          additionalProperties: { type: number }
    DecisionCriterion:
      type: object
      required: [id, name, weight, color]
      properties:
        id: { type: string }
        name: { type: string }
        weight: { type: number, minimum: 0, maximum: 100 }
        color: { type: string }
    EvidenceInput:
      type: object
      required: [title, source, claim, supportsOptionId, confidence, reliability]
      properties:
        title: { type: string, minLength: 1 }
        source: { type: string, minLength: 1 }
        url: { type: ["string", "null"] }
        claim: { type: string, minLength: 1 }
        summary: { type: string }
        supportsOptionId: { type: ["string", "null"] }
        contradictsOptionId: { type: ["string", "null"] }
        confidence: { type: number, minimum: 0, maximum: 100 }
        reliability: { type: number, minimum: 0, maximum: 100 }
    Evidence:
      allOf:
        - $ref: "#/components/schemas/EvidenceInput"
        - type: object
          required: [id, sourceType, addedBy, status]
          properties:
            id: { type: string }
            sourceType: { type: string }
            addedBy: { type: string }
            status: { type: string }
    Finding:
      type: object
      required: [id, kind, title, detail, severity]
      properties:
        id: { type: string }
        kind: { type: string }
        title: { type: string }
        detail: { type: string }
        severity: { type: string }
        evidenceIds:
          type: array
          items: { type: string }
    Recommendation:
      type: object
      required: [optionId, optionName, score, evidenceConfidence, why, whatCouldChange]
      properties:
        optionId: { type: string }
        optionName: { type: string }
        score: { type: number }
        evidenceConfidence: { type: number }
        why:
          type: array
          items: { type: string }
        whatCouldChange: { type: string }
    PendingAction:
      type: object
      required: [id, type, title, reason, status, proposedAt]
      properties:
        id: { type: string }
        type: { type: string }
        title: { type: string }
        reason: { type: string }
        status: { type: string }
        proposedAt: { type: string, format: date-time }
        currentWeights:
          type: object
          additionalProperties: { type: number }
        proposedWeights:
          type: object
          additionalProperties: { type: number }
        proposedOptionId:
          type: string
    WeightProposalInput:
      type: object
      required: [criterionId, proposedWeight, reason]
      properties:
        criterionId: { type: string }
        proposedWeight: { type: number, minimum: 0, maximum: 100 }
        reason: { type: string, minLength: 1 }
    ActionResolutionInput:
      type: object
      required: [resolution]
      properties:
        resolution:
          type: string
          enum: [approved, rejected]
    ActivityEntry:
      type: object
      required: [id, actor, action, detail, timestamp]
      properties:
        id: { type: string }
        actor: { type: string }
        action: { type: string }
        detail: { type: string }
        timestamp: { type: string, format: date-time }
    SensitivityPoint:
      type: object
      required: [weight, winner, scores]
      properties:
        weight: { type: number }
        winner: { type: string }
        scores:
          type: object
          additionalProperties: { type: number }
    SensitivityAnalysis:
      type: object
      required: [criterionName, points, stable, summary]
      properties:
        criterionName: { type: string }
        points:
          type: array
          items: { $ref: "#/components/schemas/SensitivityPoint" }
        stable: { type: boolean }
        summary: { type: string }
    DecisionBrief:
      type: object
      required: [decision, executiveRecommendation, optionsEvaluated, criteria, finalRecommendation, remainingUncertainty]
      properties:
        decision: { type: string }
        executiveRecommendation: { type: string }
        optionsEvaluated:
          type: array
          items: { type: string }
        criteria:
          type: array
          items: { type: string }
        evidenceSummary:
          type: array
          items: { type: string }
        contradictions:
          type: array
          items: { type: string }
        finalRecommendation: { type: string }
        remainingUncertainty:
          type: array
          items: { type: string }