# lib/api-zod/src/generated/types/pendingAction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ProofPilot decision intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { PendingActionCurrentWeights } from './pendingActionCurrentWeights';
import type { PendingActionProposedWeights } from './pendingActionProposedWeights';

export interface PendingAction {
  id: string;
  type: string;
  title: string;
  reason: string;
  status: string;
  proposedAt: Date;
  currentWeights?: PendingActionCurrentWeights;
  proposedWeights?: PendingActionProposedWeights;
  proposedOptionId?: string;
}